+++ HISTORY.rst
@@ -1,8 +1,13 @@
 Release History
 ===============
 
+0.0.3 (2019-05-22)
+------------------
 
-0.2.0 (2019-05-19)
+-   Do not try to download source tar.gz at first
+-   Copy files from the temp directory to avoid download again
+
+0.0.2 (2019-05-19)
 ------------------
 
 -   Birth!